Implement collapsing of weak fractional spacing and recalculation of fractional spacing.
Now when weak absolute spacing and weak fractional spacing are adjacent to each other, they collapse to one single fractional spacing with the absolute spacing as minimum.

When laying out spacings, all fractional spacings are clamped to their minimum of their share of the fractional space is to small. This converts them to absolute spacings. The remaining fractional spacings are then layed out as usual.